refactor: renames style module for intervention info components
Unifies style imports by renaming the style module to better reflect its use across intervention-related components, improving clarity and maintainability.
