+++ src/features/projectsV2/ProjectDetails/components/microComponents/InterventionHeader.tsx
@@ -1,5 +1,5 @@
 import { localizedAbbreviatedNumber } from '../../../../../utils/getFormattedNumber';
-import styles from '../../styles/PlantLocationInfo.module.scss';
+import styles from '../../styles/InterventionInfo.module.scss';
 import { useLocale, useTranslations } from 'next-intl';
 import { formatHid } from '../../../../../utils/projectV2';
