+++ src/features/projectsV2/ProjectDetails/components/microComponents/SampleTrees.tsx
@@ -2,7 +2,7 @@ import type { SetState } from '../../../../common/types/common';
 import type { SampleIntervention } from '../../../../common/types/intervention';
 
 import { useTranslations } from 'next-intl';
-import styles from '../../styles/PlantLocationInfo.module.scss';
+import styles from '../../styles/InterventionInfo.module.scss';
 
 interface Props {
   sampleInterventions: SampleIntervention[];
