+++ src/features/projectsV2/ProjectDetails/styles/InterventionInfo.module.scss
